+++ .gitignore
@@ -1,0 +1,9 @@
+venv
+* __pycache__/
+*.pyc
+.DS_Store
+logs/
+downloads/
+temp/
+*.log
+*.txt
